Select Future stale registered callback can modify state
selector future can have many callbacks registered with the channel for receiving notifications on each ask. One of the stale callback getting executed while closing the channel(after getting a value on the future) can modify the current state that indicates that we have processed this future for the select. This with batch replay can end loosing notifying select.