Add type to use as keys for values attached to the BackgroundActivityContext
go lint rules error out if the type of the key passed to context.WithValue
is a basic type. The official golang documentation for context also provides
an example of using strongly typed keys: https://blog.golang.org/context#TOC_3.2.

This change provides a type in the Cadence client library that can be used to
attach data to the BackgroundActivityContext.