Un-pin zap, bump to a newer version to pull in zaptest logger
Mostly motivated by the addition of a testing.T.Logf-compatible logger
provided by zap, which allows quite a lot cleaner test-logs, especially
when they are run in parallel.

But generally: there's no reason for zap to be pinned to a *specific*
version like it was. Ranges are far nicer for users of libraries.