Add ResetStickyTaskList API to cadence client
This commit adds ResetStickyTaskList API to cadence client, which task
handlers can use to communicate to the server not to schedule any
further decision tasks onto the task handler's private sticky tasklist.

One use case would be when a decision worker evicts explicitly a workflow
context from its workflow cache, for example due to running out of cache
slots, it should call ResetStickyTaskList to indicate to the server there
would be no advantage in putting further tasks of this workflow onto
the sticky tasklist because the workflow would have to be replayed from
the start anyways due to having been evicted from cache.