Improve error when workflow is missing arguments to an activity
Previously it would panic with a json EOF error without any indication of where that happened.
An even better way would be if it was possible to get the stacktrace from the workflow execution rather than the activity execution, but that didn't seem to be possible in any easy way.