Fix binary checksum race condition and lack of err handling
In practice this is probably only an issue in tests, as other use of the checksum
depends on a running worker which would have initialized the value.
Still, if it's un-protected, Go has no obligation to propagate the change, so it's
risky to leave.

This now also detects and panics on errs from lazily initializing the checksum,
and explicitly ignores the Close() error (to keep errcheck happy, if used) since
it's relatively safe to do so.