encoded: Unalias interfaces
Per reasoning in #928, this redeclares all three interfaces in the
`encoded` package and adds bi-directional compile-time checks to verify
that the interfaces do not fall out-of-sync.

The only reason I could find for the interfaces to be aliased was
`client.Client`'s re-use of an `internal.Client`.

    package client

    type Client interface {
      ...
      QueryWorkflow(...) (encoded.Value, error)
    }

    func NewClient(...) Client {
      return internal.NewClient(...)
    }

Where,

    package internal

    type Client interface {
      ...
      QueryWorkflow(...) (internal.Value, error)
    }

This mismatch in `QueryWorkflow`'s return value makes `internal.Client`
not match `client.Client`.

To resolve this, we introduce a `clientAdapter` which adapts an
`internal.Client` into a `client.Client`. It does so by embedding
`internal.Client` and implementing the compliant version of
`QueryWorkflow`.

Resolves #928