Backwards compatible activity type names
Previously merged shortened activity type names of form {Method} is not
compatible with current users of this client, since multiple activities
in different packages with the same name results in naming conflicts.

This change reverts shortening when no activity alias (prefix) is provided.
The use of prefix shortens the name, assuming the users responsiblity to
choose proper prefix.