Use test-scoped loggers everywhere
... or nearly.  TestWorkerOptionDefaults currently produces a log because
we have an info-level log when no logger is set, and... I think that makes sense.
```
=== RUN   TestWorkerOptionDefaults
{"level":"info","ts":"2021-10-11T16:42:58.778-0700","caller":"internal/internal_worker.go:223","msg":"No logger configured for cadence worker. Created default one."}
{"level":"info","ts":"2021-10-11T16:42:58.779-0700","caller":"internal/internal_worker.go:227","msg":"No metrics scope configured for cadence worker. Use NoopScope as default."}
```

A better behavior may be to wait to produce that log until the worker is started,
but that's a bit out of scope for this PR + we may not want to do that anyway.