Creating cadence.IsWorkflowError helper (#1145)
We have a variety of from-workflow errors that are relevant for users calling
`client.ExecuteWorkflow(...).Get(...)` and `client.GetWorkflow(...).Get(...)`.

Since they do not share any common wrapper or "parent type" that would allow
easy comparison with an error type check or `errors.As`, they are rather
annoying to differentiate from general RPC errors like timeouts or connection
failures.

So this adds a top-level "check all relevant types" helper, which is
intended to be a stable target for users to rely on, even if we add more
types later.  It intentionally does not tell *which* kind of error it is, in
case there are multiple types in the wrapped chain, as then behavior would
be order-dependent and any change risks breaking someone's code.