Hopeful-mitigation for concurrent map writes on goroutine shutdown
Workflow goroutines are shut down concurrently, which can lead to race
conditions and crashes.  This is an attempt to make them less likely
to terminate the hosting process, specifically due to concurrent map
writes and reads.

We *unambiguously* need to fix the underlying issue, but that is
substantially more complicated and we're unfortunately pretty busy.

---

1ms is likely overkill, but even 1us is almost certainly insufficient.
There can be non-trivial behavior in deferred functions.

Still, this 1ms is a wild guess, it can be changed or made configurable
and it'll likely be fine.