Maintain a stable order of children context, hopefully resolves some non-determinism around cancels
We have some non-deterministic failures in a user's workflows, which (as far as either of us can tell)
are not due to flawed workflow code.  We have been unable to reliably reproduce it to investigate, but
there is this map of contexts which are iterated over to propagate cancellation information, and that
seems at least slightly suspect.

I'm not confident that this map can or cannot cause non-determinism elsewhere, but making it clearly
deterministic seems both safe[1] and reduces questionable code.

---

I've opted to keep the list in order while removing entries because it's likely to be a small list (and
even if it is not, it'll be dwarfed by other costs associated with code that caused it to be large),
and because it seems probably easier to reason about by hand if needed: the list of cancellation signals
will be first-in, first-out, rather than being based on when things are removed from the list.

Either way should be perfectly repeatable though, so I'm happy to be convinced to do the more efficient
thing instead.  I'm not totally sold on its utility anyway.

---

[1]: If the map's random order does not cause non-determinism failures, changing the execution order
     like this PR does will also not cause non-determinism failures.
     If it *does*, then this change is no worse (it will not introduce *more* non-determinism due to
     the change) and it may resolve the problem for future workflows.
     So this is at worst identical, and at best a fix.