Honor non-determinism fail workflow policy (#1287)
What changed?
Users can specify NonDeterministicWorkflowPolicy in worker options. If the FailWorkflow policy is chosen the workflow is expected to terminate as soon as it ends up with a nondeterministic state (e.g. activity order changed).
However this wasn't honored for a category of nondeterminism cases. This PR addresses it and workflows fail once any nondeterminism scenario is encountered.

There are two categories of nondeterminism cases in terms of how they get detected by client library:
1. Issue bubbles up as illegal state panic to the task handler. Most actual prod cases.
2. Issue is caught when comparing replay decisions with history. Replay test scenarios and a subset of prod cases.

FailWorkflow policy was honored for 2 but not for 1.

Why?
To make NonDeterministicWorkflowPolicy feature correct/complete.

How did you test it?
Added an integration test to simulate this scenario.